refactor: internal modules live in src/internal/, not _-prefixed files
_esbuild, _snapshot, _test-browser, and _test-runner move to
src/internal/{esbuild,snapshot,test-browser,test-runner}.ts. A
subdirectory already hides modules from entrypoint discovery (only
top-level src files are considered - src/plugins/ has relied on this
all along), so the directory IS the convention and the underscore
prefix was carrying the same information redundantly.

Entrypoints after the move: cli, libuild, test - unchanged surface
(verified against dist/package.json exports). dist gains internal/
for the relocated declarations (#11); runtime code chunks as before.
Consumer e2e re-verified on bun, node, and chromium.
