+++ src/cli.ts
@@ -3,7 +3,7 @@ import { Command } from "commander";
 import * as FS from "fs/promises";
 import * as Path from "path";
 import { build, publish, stage } from "./libuild.ts";
-import { resolveTestTargets, runTests, type Platform } from "./_test-runner.ts";
+import { resolveTestTargets, runTests, type Platform } from "./internal/test-runner.ts";
 
 // =============================================================================
 // Publish argument validation
